Update plugin version and fix attribute assignment
The commit includes an increment in the 'Author Avatars' plugin version to 2.1.20. It also rectifies attribute assignment within the 'display' array in the show-avatar functionality. The correction ensures the key rather than the value is added to the 'display' array, enhancing accuracy and functionality. The difference is also recorded in the readme.txt's changelog section.